requirements.txt
flask
python-dotenv
openai
gunicorn
flask-cors